+++ src/Communications/KShop.Communications.Contracts/Products/ProductsReserveMap.cs
@@ -6,13 +6,13 @@ namespace KShop.Communications.Contracts.Products
     /// Показывает какие ID зарезервированных позиций соответствуют ID продукта
     /// [product_id, reserve_id]
     /// </summary>
-    public class ProductsReserveMap : Dictionary<ulong, ulong>
+    public class ProductsReserveMap : Dictionary<uint, uint>
     {
         public ProductsReserveMap()
         {
         }
 
-        public ProductsReserveMap(IDictionary<ulong, ulong> dictionary) : base(dictionary)
+        public ProductsReserveMap(IDictionary<uint, uint> dictionary) : base(dictionary)
         {
         }
     }
