+++ src/Orders/KShop.Orders.Persistence/Entities/OrderPosition.cs
@@ -4,9 +4,9 @@ namespace KShop.Orders.Persistence.Entities
 {
     public class OrderPosition
     {
-        public ulong ID { get; set; }
+        public uint ID { get; set; }
         public Guid OrderID { get; set; }
-        public ulong ProductID { get; set; }
+        public uint ProductID { get; set; }
         public uint Quantity { get; set; }
